+++ changelog.md
@@ -6,6 +6,7 @@ Version 19.03
 
 + String `__version__` provided.
 + Triangulation data type changed for performance improvement.
++ Change "print" function as "logger.debug".
 
 Version 19.02
 ===